+++ packages/plasma-temple/src/components/Cart/Cart@sberportal.tsx
@@ -1,0 +1,14 @@
+import React from 'react';
+
+import { CartCommon, CartComponent } from './Cart@common';
+import { CartItemListCommon } from './CartItemList/CartItemList@common';
+import { CartOrderSberPortal } from './CartOrder/CartOrder@sberportal';
+
+const platformComponents = {
+    CartItemList: CartItemListCommon,
+    CartOrder: CartOrderSberPortal,
+};
+
+export const CartSberPortal: CartComponent = (props) => (
+    <CartCommon {...props} platformComponents={platformComponents} />
+);
