+++ packages/plasma-temple/src/components/Cart/CartItemList/CartItemList.tsx
@@ -1,16 +1,16 @@
 import React from 'react';
-import { DeviceKind } from '@sberdevices/plasma-ui';
 
-import { deviceFamily } from '../../../utils/deviceFamily';
+import { DeviceComponent } from '../../DeviceComponent/DeviceComponent';
 
 import { CartItemListCommon, CartItemListProps } from './CartItemList@common';
 import { CartItemListMobile } from './CartItemList@mobile';
 
-export const mapDeviceToCartItem: Record<DeviceKind, React.FC<CartItemListProps>> = {
-    sberBox: CartItemListCommon,
-    sberPortal: CartItemListCommon,
-    mobile: CartItemListMobile,
-};
-
 /** Компонент отображает товары в корзине  */
-export const CartItemList = mapDeviceToCartItem[deviceFamily] as typeof CartItemListCommon;
+export const CartItemList: typeof CartItemListCommon = (props) => (
+    <DeviceComponent
+        sberbox={CartItemListCommon}
+        sberportal={CartItemListCommon}
+        mobile={CartItemListMobile}
+        props={props as CartItemListProps}
+    />
+);
