+++ packages/plasma-temple/src/components/Cart/CartOrder/CartOrder@sberportal.tsx
@@ -3,7 +3,7 @@ import styled from 'styled-components';
 import { Button, Footnote1 } from '@sberdevices/plasma-ui';
 import { secondary } from '@sberdevices/plasma-tokens';
 
-import { CartOrderDetails } from '../CartOrderDetails/CartOrderDetails';
+import { CartOrderDetailsCommon } from '../CartOrderDetails/CartOrderDetails@common';
 import { AnyObject } from '../../../types';
 
 import { CartOrderProps, CheckoutButtonContent } from './CartOrder@common';
@@ -14,7 +14,7 @@ const StyledContainer = styled.div`
     margin-top: 0.375rem;
 `;
 
-const StyledCartOrderDetails = styled(CartOrderDetails)`
+const StyledCartOrderDetails = styled(CartOrderDetailsCommon)`
     margin-bottom: 1.5rem;
 `;
