+++ packages/plasma-temple/src/components/Cart/index@mobile.ts
@@ -1,0 +1,4 @@
+export { CartMobile as Cart } from './Cart@mobile';
+export { CartItemListMobile as CartItemList } from './CartItemList/CartItemList@mobile';
+export { CartOrderMobile as CartOrder } from './CartOrder/CartOrder@mobile';
+export { EmptyCartMobile as EmptyCart } from './EmptyCart/EmptyCart@mobile';
