+++ packages/plasma-temple/src/components/Cart/index@sberbox.ts
@@ -1,0 +1,4 @@
+export { CartSberBox as Cart } from './Cart@sberbox';
+export { CartItemListCommon as CartItemList } from './CartItemList/CartItemList@common';
+export { CartOrderSberBox as CartOrder } from './CartOrder/CartOrder@sberbox';
+export { EmptyCartCommon as EmptyCart } from './EmptyCart/EmptyCart@common';
