+++ packages/plasma-temple/src/components/Cart/index@sberportal.ts
@@ -1,0 +1,4 @@
+export { CartSberPortal as Cart } from './Cart@sberportal';
+export { CartItemListCommon as CartItemList } from './CartItemList/CartItemList@common';
+export { CartOrderSberPortal as CartOrder } from './CartOrder/CartOrder@sberportal';
+export { EmptyCartCommon as EmptyCart } from './EmptyCart/EmptyCart@common';
