+++ packages/plasma-temple/src/components/index@mobile.ts
@@ -1,2 +1,3 @@
 export { StateLayoutMobile as StateLayout } from './StateLayout/StateLayout@mobile';
 export * from './Item/index@mobile';
+export * from './Cart/index@mobile';
