+++ packages/plasma-temple/src/components/index@sberbox.ts
@@ -1,2 +1,3 @@
 export { StateLayoutSberBox as StateLayout } from './StateLayout/StateLayout@sberbox';
 export * from './Item/index@sberbox';
+export * from './Cart/index@sberbox';
