+++ packages/plasma-temple/src/components/index@sberportal.ts
@@ -1,2 +1,3 @@
 export { StateLayoutSberPortal as StateLayout } from './StateLayout/StateLayout@sberportal';
 export * from './Item/index@sberportal';
+export * from './Cart/index@sberportal';
